:hammer: Move core functionality to InitPasteDataService for reuse across multiple platforms
