+++ app/src/commonMain/kotlin/com/crosspaste/app/AppLaunchState.kt
@@ -1,3 +1,8 @@
 package com.crosspaste.app
 
-interface AppLaunchState
+interface AppLaunchState {
+
+    val acquireLock: Boolean
+
+    val firstLaunch: Boolean
+}
