+++ app/src/desktopMain/kotlin/com/crosspaste/app/DesktopAppLaunchState.kt
@@ -2,8 +2,8 @@ package com.crosspaste.app
 
 data class DesktopAppLaunchState(
     val pid: Long,
-    val acquireLock: Boolean,
-    val firstLaunch: Boolean,
+    override val acquireLock: Boolean,
+    override val firstLaunch: Boolean,
     var accessibilityPermissions: Boolean,
     var installFrom: String?,
 ) : AppLaunchState
